% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastamp-package.R
\docType{package}
\name{datastamp-package}
\alias{datastamp}
\alias{datastamp-package}
\title{datastamp: Stamping data with metadata}
\description{
The goal of datastamp is to make it easy to attach some metadata to 
    R objects. This can be convenient if you want to make `.RData` or `.rds` 
    objects self-documenting, by attaching time of creation or the path to the 
    script used to generate the data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/teunbrand/datastamp}
  \item Report bugs at \url{https://github.com/teunbrand/datastamp/issues}
}

}
\author{
\strong{Maintainer}: Teun van den Brand \email{tahvdbrand@gmail.com} (\href{https://orcid.org/0000-0002-9335-7468}{ORCID})

}
\keyword{internal}
