% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_document.R
\name{get_current_document}
\alias{get_current_document}
\title{Guess the current document}
\usage{
get_current_document()
}
\value{
A \code{character} with the path to the document
}
\description{
\code{get_current_document()} makes a guess at what the path to the current
document is. Returns the character \code{"console"} when it can't find the
document. This is a decent proxy for finding the document this function was
run in.
}
\details{
When executed interactively within RStudio, it reports the currently
opened document at the time of execution. If the active document is
switched between queuing this function and execution, the document of
origin might be misreported.
}
\examples{
doc <- get_current_document()
}
