% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{get_stamp}
\alias{get_stamp}
\title{Retrieve stamp}
\usage{
get_stamp(object)
}
\arguments{
\item{object}{An object.}
}
\value{
A \code{datastamp} object.
}
\description{
If an object contains a datastamp, returns the datastamp.
}
\examples{
x <- stamp(1:5)

get_stamp(x)
}
