% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info_stamp}
\alias{info_stamp}
\alias{info_script}
\alias{info_time}
\alias{info_session}
\alias{info_system}
\alias{info_user}
\alias{info_version}
\alias{info_files}
\title{Retrieve information from a datastamp}
\usage{
info_script(object)

info_time(object)

info_session(object)

info_system(object)

info_user(object)

info_version(object)

info_files(object)
}
\arguments{
\item{object}{An object to retrieve information from.}
}
\value{
Either the requested information or \code{NULL} when the information
couldn't be retrieved.
}
\description{
These are convenience function for extracting information from a
datastamp or object with a datastamp.
}
\examples{
x <- stamp(data.frame(x = 1:5, y = 5:1))

info_script(x)
info_time(x)
info_session(x)
info_system(x)
info_user(x)
info_version(x)
info_files(x)
}
