% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{is_stamp}
\alias{is_stamp}
\title{Type test for datastamp objects.}
\usage{
is_stamp(object)
}
\arguments{
\item{object}{An object.}
}
\value{
A \code{logical} of length 1.
}
\description{
Test for objects of type \code{datastamp}.
}
\examples{
obj <- stamp(1:5)

is_stamp(obj) # FALSE

is_stamp(get_stamp(obj)) # TRUE
}
