% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{make_stamp}
\alias{make_stamp}
\title{Build a datastamp}
\usage{
make_stamp(
  script = NULL,
  time = NULL,
  system = NULL,
  session = NULL,
  files = NULL
)
}
\arguments{
\item{script}{A \code{logical} to include the likely script of origin (default:
\code{TRUE}).}

\item{time}{A \code{logical} to include a timestamp in the datastamp (default:
\code{TRUE}).}

\item{system}{A \code{logical} to include system information (default: \code{TRUE}).}

\item{session}{A \code{logical} to include session information (default: \code{TRUE})}

\item{files}{A \code{character} vector with file paths.}
}
\value{
A \code{list} of the class \code{datastamp}.
}
\description{
Searches for relevant data to make a datastamp.
}
\details{
When the \code{script}, \code{time}, \code{system} or \code{session} arguments are
\code{NULL} the global options are searched for \code{"datastamp.*"} options, wherein
the \code{*} is replaced by the argument names.
}
\examples{
# Default stamp
stamp <- make_stamp()

# Not including some information:
make_stamp(session = FALSE)

# Alternatively:
options("datastamp.session" = FALSE)
make_stamp()
}
