% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{set_stamp}
\alias{set_stamp}
\alias{set_stamp.default}
\alias{set_stamp.datastamp}
\alias{set_stamp.Vector}
\alias{get_stamp.default}
\alias{get_stamp.Vector}
\alias{get_stamp.datastamp}
\title{Stamp setter}
\usage{
set_stamp(object, stamp)

\method{set_stamp}{default}(object, stamp)

\method{set_stamp}{datastamp}(object, stamp)

\method{set_stamp}{Vector}(object, stamp)

\method{get_stamp}{default}(object)

\method{get_stamp}{Vector}(object)

\method{get_stamp}{datastamp}(object)
}
\arguments{
\item{object}{An R object.}

\item{stamp}{A \code{datastamp} object.}
}
\value{
The \code{object} with a datastamp attached.
}
\description{
Attaches a datastamp to an object.
}
\details{
For Bioconductor objects in the S4Vectors framework, the datastamp
is stored in the \code{metadata} slot instead of the attributes.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Attaches \code{stamp} to \code{datastamp} attribute.

\item \code{datastamp}: Raises an error.

\item \code{Vector}: Attaches \code{stamp} to list in \code{metadata} slot.

\item \code{default}: Retrieves the \code{"datastamp"} attribute.

\item \code{Vector}: Retrieves the list-item named \code{"datastamp"} form the
\code{metadata} slot.

\item \code{datastamp}: Returns \code{object} itself.
}}

\examples{
x <- 1:5
y <- make_stamp()

x <- set_stamp(x, y)
}
